+++ test/unit/bluesky/extrafilewriters/test_firescsvs.py
@@ -100,7 +100,8 @@ class TestFiresCsvsWriterCollectCsvFields(object):
                                 {
                                     "lat": 35.0,
                                     "lng": -96.2,
-                                    "area": 99
+                                    "area": 99,
+                                    "source": "GOES-16"
                                 }
                             ]
                         },
@@ -173,6 +174,7 @@ class TestFiresCsvsWriterCollectCsvFields(object):
 
         expected_fires_fields = [
             {
+                "source": "GOES-16",
                 "area": 99,
                 "canopy_consumption_pct": '',
                 'ch4': '',
@@ -189,6 +191,7 @@ class TestFiresCsvsWriterCollectCsvFields(object):
                 'event_id': '',
                 'event_name': '',
                 'fccs_number': '',
+                'fuelbed_fractions': '',
                 'heat': '',
                 'id': 'SF11C14225236095807750',
                 "latitude": 35.0,
@@ -225,6 +228,7 @@ class TestFiresCsvsWriterCollectCsvFields(object):
                 'voc': ''
             },
             {
+                "source": "",
                 "area": 102,
                 "canopy_consumption_pct": 23.3,
                 'ch4': '',
@@ -241,6 +245,7 @@ class TestFiresCsvsWriterCollectCsvFields(object):
                 'event_id': '',
                 'event_name': '',
                 'fccs_number': '',
+                'fuelbed_fractions': '',
                 'heat': '',
                 'id': 'SF11C14225236095807750',
                 "latitude": 30.0,
@@ -277,6 +282,7 @@ class TestFiresCsvsWriterCollectCsvFields(object):
                 'voc': ''
             },
             {
+                'source': "",
                 'area': 120.0,
                 "canopy_consumption_pct": 23.3,
                 'ch4': '',
@@ -293,6 +299,7 @@ class TestFiresCsvsWriterCollectCsvFields(object):
                 'event_id': '',
                 'event_name': '',
                 'fccs_number': '9',
+                'fuelbed_fractions': '9 1.0',
                 'heat': 64442794218.7609,
                 'id': 'SF11C14225236095807750',
                 'latitude': 47.41,
@@ -347,7 +354,7 @@ class TestFiresCsvsWriterCollectCsvFields(object):
                 "id": "SF11C14225236095807750",
                 "event_of": {
                     "id": "bigeventid"
-                    # porposely missing 'name'
+                    # purposely missing 'name'
                 },
                 "type": "wildfire",
                 "activity": [
@@ -444,6 +451,74 @@ class TestFiresCsvsWriterCollectCsvFields(object):
                                         ],
                                         "lat": 47.41,
                                         "lng": -121.41
+                                    },
+                                    {
+                                        "area": 120.0,
+                                        "rain_days": 8,
+                                        "slope": 20.0,
+                                        "snow_month": 5,
+                                        "sunrise_hour": 4,
+                                        "sunset_hour": 19,
+                                        "consumption": {
+                                            "summary": {
+                                                "flaming": 1311.2071801109494,
+                                                "residual": 1449.3962581338644,
+                                                "smoldering": 1267.0712004277434,
+                                                "total": 4027.6746386725567
+                                            }
+                                        },
+                                        "fuelbeds": [
+                                            {
+                                                "emissions": {
+                                                    "flaming": {
+                                                        "PM2.5": [9.545588271207714]
+                                                    },
+                                                    "residual": {
+                                                        "PM2.5": [24.10635856528243]
+                                                    },
+                                                    "smoldering": {
+                                                        "PM2.5": [21.073928205514225]
+                                                    },
+                                                    "total": {
+                                                        "PM2.5": [54.725875042004375]
+                                                    }
+                                                },
+                                                "fccs_id": "9",
+                                                "heat": {
+                                                    "flaming": [20979314881.77519],
+                                                    "residual": [23190340130.141827],
+                                                    "smoldering": [20273139206.843895],
+                                                    "total": [64442794218.7609]
+                                                },
+                                                "pct": 40.0
+                                            },
+                                            {
+                                                "emissions": {
+                                                    "flaming": {
+                                                        "PM2.5": [0]
+                                                    },
+                                                    "residual": {
+                                                        "PM2.5": [0]
+                                                    },
+                                                    "smoldering": {
+                                                        "PM2.5": [0]
+                                                    },
+                                                    "total": {
+                                                        "PM2.5": [10]
+                                                    }
+                                                },
+                                                "fccs_id": "52",
+                                                "heat": {
+                                                    "flaming": [0],
+                                                    "residual": [0],
+                                                    "smoldering": [0],
+                                                    "total": [0]
+                                                },
+                                                "pct": 60.0
+                                            }
+                                        ],
+                                        "lat": 47.41,
+                                        "lng": -121.41
                                     }
                                 ],
                                 "state": "WA"
@@ -460,6 +535,7 @@ class TestFiresCsvsWriterCollectCsvFields(object):
 
         expected_fires_fields = [
             {
+                "source": "",
                 "area": 99,
                 "canopy_consumption_pct": '',
                 'ch4': '',
@@ -476,6 +552,7 @@ class TestFiresCsvsWriterCollectCsvFields(object):
                 'event_id': 'bigeventid',
                 'event_name': '',
                 'fccs_number': '',
+                'fuelbed_fractions': '',
                 'heat': '',
                 'id': 'SF11C14225236095807750',
                 "latitude": 35.0,
@@ -512,6 +589,7 @@ class TestFiresCsvsWriterCollectCsvFields(object):
                 'voc': ''
             },
             {
+                "source": "",
                 "area": 102,
                 "canopy_consumption_pct": 23.3,
                 'ch4': '',
@@ -528,6 +606,7 @@ class TestFiresCsvsWriterCollectCsvFields(object):
                 'event_id': 'bigeventid',
                 'event_name': 'big event name',
                 'fccs_number': '',
+                'fuelbed_fractions': '',
                 'heat': '',
                 'id': 'SF11C14225236095807750-B',
                 "latitude": 30.0,
@@ -564,6 +643,7 @@ class TestFiresCsvsWriterCollectCsvFields(object):
                 'voc': ''
             },
             {
+                'source': "",
                 'area': 120.0,
                 "canopy_consumption_pct": 23.3,
                 'ch4': '',
@@ -580,6 +660,7 @@ class TestFiresCsvsWriterCollectCsvFields(object):
                 'event_id': 'bigeventid',
                 'event_name': 'big event name',
                 'fccs_number': '9',
+                'fuelbed_fractions': '9 1.0',
                 'heat': 64442794218.7609,
                 'id': 'SF11C14225236095807750-B',
                 'latitude': 47.41,
@@ -614,6 +695,60 @@ class TestFiresCsvsWriterCollectCsvFields(object):
                 'type': 'WF',
                 'utc_offset': '-07:00',
                 'voc': ''
+            },
+            {
+                'source': "",
+                'area': 120.0,
+                "canopy_consumption_pct": 23.3,
+                'ch4': '',
+                'co': '',
+                'co2': '',
+                'consumption_flaming': 1311.2071801109494,
+                'consumption_residual': 1449.3962581338644,
+                'consumption_smoldering': 1267.0712004277434,
+                'consumption_total': 4027.6746386725567,
+                'country': '',
+                'county': '',
+                'date_time': '20150804',
+                'elevation': '',
+                'event_id': 'bigeventid',
+                'event_name': 'big event name',
+                'fccs_number': '52',
+                'fuelbed_fractions': '52 0.6; 9 0.4',
+                'heat': 64442794218.7609,
+                'id': 'SF11C14225236095807750-B',
+                'latitude': 47.41,
+                'longitude': -121.41,
+                'max_humid': '',
+                'max_temp': '',
+                'max_temp_hour': '',
+                'max_wind': '',
+                'max_wind_aloft': '',
+                'min_humid': '',
+                'min_temp': '',
+                'min_temp_hour': '',
+                'min_wind': 34,
+                'min_wind_aloft': '',
+                'moisture_100hr': '',
+                'moisture_10hr': '',
+                'moisture_1hr': '',
+                'moisture_1khr': '',
+                'moisture_duff': '',
+                'moisture_live': '',
+                'nh3': '',
+                'nox': '',
+                'pm10': '',
+                'pm2.5': 64.725875042004375,
+                'rain_days': 8.0,
+                'slope': 20.0,
+                'snow_month': 5.0,
+                'so2': '',
+                'state': 'WA',
+                'sunrise_hour': 4.0,
+                'sunset_hour': 19.0,
+                'type': 'WF',
+                'utc_offset': '-07:00',
+                'voc': ''
             }
         ]
 
@@ -626,7 +761,7 @@ class TestFiresCsvsWriterCollectCsvFields(object):
         expected_events_fields = {
             'bigeventid': {
                 'name':'big event name',
-                'total_area': 321.0,
+                'total_area': 441.0,
                 'total_ch4': None,
                 'total_co': None,
                 'total_co2': None,
@@ -635,7 +770,381 @@ class TestFiresCsvsWriterCollectCsvFields(object):
                 'total_nmhc': None,
                 'total_nox': None,
                 'total_pm10': None,
-                'total_pm2.5': 64.725875042004375,
+                'total_pm2.5': None, # because pm2.5 wasn't defined for one location
+                'total_so2': None,
+                'total_voc': None
+            }
+        }
+        assert events_fields == expected_events_fields
+
+    def test_two_fires_one_event_two_days(self):
+        fires = [
+            Fire({
+                "fuel_type": "natural",
+                "id": "SF11C14225236095807750",
+                "event_of": {
+                    "id": "bigeventid"
+                    # purposely missing 'name'
+                },
+                "type": "wildfire",
+                # two days defined as two active areas in the same
+                # actividy object
+                "activity": [
+                    {
+                        "active_areas": [
+                            {
+                                "start": "2015-08-04T18:00:00",
+                                "end": "2015-08-05T18:00:00",
+                                "utc_offset": "-06:00",
+                                "specified_points": [
+                                    {
+                                        "lat": 35.0,
+                                        "lng": -96.2,
+                                        "area": 99,
+                                        "fuelbeds": [
+                                            {
+                                                "emissions": {
+                                                    "total": {
+                                                        "PM2.5": [10]
+                                                    }
+                                                }
+                                            }
+                                        ]
+                                    }
+                                ]
+                            },
+                            {
+                                "start": "2015-08-05T18:00:00",
+                                "end": "2015-08-06T18:00:00",
+                                "utc_offset": "-06:00",
+                                "specified_points": [
+                                    {
+                                        "lat": 35.0,
+                                        "lng": -96.2,
+                                        "area": 99,
+                                        "fuelbeds": [
+                                            {
+                                                "emissions": {
+                                                    "total": {
+                                                        "PM2.5": [20]
+                                                    }
+                                                }
+                                            }
+                                        ]
+                                    }
+                                ]
+                            },
+                        ]
+                    }
+                ]
+            }),
+            Fire({
+                "fuel_type": "natural",
+                "id": "SF11C14225236095807750",
+                "event_of": {
+                    "id": "bigeventid",
+                    "name": "foo-event"
+                },
+                "type": "wildfire",
+                # two days defined as two activity objects, each with
+                # one active area
+                "activity": [
+                    {
+                        "active_areas": [
+                            {
+                                "start": "2015-08-04T18:00:00",
+                                "end": "2015-08-05T18:00:00",
+                                "utc_offset": "-06:00",
+                                "specified_points": [
+                                    {
+                                        "lat": 35.5,
+                                        "lng": -96.0,
+                                        "area": 200,
+                                        "fuelbeds": [
+                                            {
+                                                "emissions": {
+                                                    "total": {
+                                                        "PM2.5": [30]
+                                                    }
+                                                }
+                                            }
+                                        ]
+                                    }
+                                ]
+                            }
+                        ]
+                    },
+                    {
+                        "active_areas": [
+                            {
+                                "start": "2015-08-05T18:00:00",
+                                "end": "2015-08-06T18:00:00",
+                                "utc_offset": "-06:00",
+                                "specified_points": [
+                                    {
+                                        "lat": 35.5,
+                                        "lng": -96.0,
+                                        "area": 200,
+                                        "fuelbeds": [
+                                            {
+                                                "emissions": {
+                                                    "total": {
+                                                        "PM2.5": [40]
+                                                    }
+                                                }
+                                            }
+                                        ]
+                                    }
+                                ]
+                            },
+                        ]
+                    }
+                ]
+            })
+        ]
+
+
+        writer = firescsvs.FiresCsvsWriter('/foo')
+        fires_fields, events_fields = writer._collect_csv_fields(fires)
+
+        expected_fires_fields = [
+            {
+                "source": "",
+                "area": 99,
+                "canopy_consumption_pct": '',
+                'ch4': '',
+                'co': '',
+                'co2': '',
+                'consumption_flaming': '',
+                'consumption_residual': '',
+                'consumption_smoldering': '',
+                'consumption_total': '',
+                'country': '',
+                'county': '',
+                'date_time': '20150804',
+                'elevation': '',
+                'event_id': 'bigeventid',
+                'event_name': '',
+                'fccs_number': '',
+                'fuelbed_fractions': '',
+                'heat': '',
+                'id': 'SF11C14225236095807750',
+                "latitude": 35.0,
+                "longitude": -96.2,
+                'max_humid': '',
+                'max_temp': '',
+                'max_temp_hour': '',
+                'max_wind': '',
+                'max_wind_aloft': '',
+                'min_humid': '',
+                'min_temp': '',
+                'min_temp_hour': '',
+                'min_wind': '',
+                'min_wind_aloft': '',
+                'moisture_100hr': '',
+                'moisture_10hr': '',
+                'moisture_1hr': '',
+                'moisture_1khr': '',
+                'moisture_duff': '',
+                'moisture_live': '',
+                'nh3': '',
+                'nox': '',
+                'pm10': '',
+                'pm2.5': 10.0,
+                'rain_days': '',
+                'slope': '',
+                'snow_month': '',
+                'so2': '',
+                'state': '',
+                'sunrise_hour': '',
+                'sunset_hour': '',
+                'type': 'WF',
+                "utc_offset": "-06:00",
+                'voc': ''
+            },
+            {
+                "source": "",
+                "area": 99,
+                "canopy_consumption_pct": '',
+                'ch4': '',
+                'co': '',
+                'co2': '',
+                'consumption_flaming': '',
+                'consumption_residual': '',
+                'consumption_smoldering': '',
+                'consumption_total': '',
+                'country': '',
+                'county': '',
+                'date_time': '20150805',
+                'elevation': '',
+                'event_id': 'bigeventid',
+                'event_name': '',
+                'fccs_number': '',
+                'fuelbed_fractions': '',
+                'heat': '',
+                'id': 'SF11C14225236095807750',
+                "latitude": 35.0,
+                "longitude": -96.2,
+                'max_humid': '',
+                'max_temp': '',
+                'max_temp_hour': '',
+                'max_wind': '',
+                'max_wind_aloft': '',
+                'min_humid': '',
+                'min_temp': '',
+                'min_temp_hour': '',
+                'min_wind': '',
+                'min_wind_aloft': '',
+                'moisture_100hr': '',
+                'moisture_10hr': '',
+                'moisture_1hr': '',
+                'moisture_1khr': '',
+                'moisture_duff': '',
+                'moisture_live': '',
+                'nh3': '',
+                'nox': '',
+                'pm10': '',
+                'pm2.5': 20.0,
+                'rain_days': '',
+                'slope': '',
+                'snow_month': '',
+                'so2': '',
+                'state': '',
+                'sunrise_hour': '',
+                'sunset_hour': '',
+                'type': 'WF',
+                "utc_offset": "-06:00",
+                'voc': ''
+            },
+            {
+                "source": "",
+                "area": 200,
+                "canopy_consumption_pct": '',
+                'ch4': '',
+                'co': '',
+                'co2': '',
+                'consumption_flaming': '',
+                'consumption_residual': '',
+                'consumption_smoldering': '',
+                'consumption_total': '',
+                'country': '',
+                'county': '',
+                'date_time': '20150804',
+                'elevation': '',
+                'event_id': 'bigeventid',
+                'event_name': 'foo-event',
+                'fccs_number': '',
+                'fuelbed_fractions': '',
+                'heat': '',
+                'id': 'SF11C14225236095807750',
+                "latitude": 35.5,
+                "longitude": -96.0,
+                'max_humid': '',
+                'max_temp': '',
+                'max_temp_hour': '',
+                'max_wind': '',
+                'max_wind_aloft': '',
+                'min_humid': '',
+                'min_temp': '',
+                'min_temp_hour': '',
+                'min_wind': '',
+                'min_wind_aloft': '',
+                'moisture_100hr': '',
+                'moisture_10hr': '',
+                'moisture_1hr': '',
+                'moisture_1khr': '',
+                'moisture_duff': '',
+                'moisture_live': '',
+                'nh3': '',
+                'nox': '',
+                'pm10': '',
+                'pm2.5': 30.0,
+                'rain_days': '',
+                'slope': '',
+                'snow_month': '',
+                'so2': '',
+                'state': '',
+                'sunrise_hour': '',
+                'sunset_hour': '',
+                'type': 'WF',
+                "utc_offset": "-06:00",
+                'voc': ''
+            },
+            {
+                "source": "",
+                "area": 200,
+                "canopy_consumption_pct": '',
+                'ch4': '',
+                'co': '',
+                'co2': '',
+                'consumption_flaming': '',
+                'consumption_residual': '',
+                'consumption_smoldering': '',
+                'consumption_total': '',
+                'country': '',
+                'county': '',
+                'date_time': '20150805',
+                'elevation': '',
+                'event_id': 'bigeventid',
+                'event_name': 'foo-event',
+                'fccs_number': '',
+                'fuelbed_fractions': '',
+                'heat': '',
+                'id': 'SF11C14225236095807750',
+                "latitude": 35.5,
+                "longitude": -96.0,
+                'max_humid': '',
+                'max_temp': '',
+                'max_temp_hour': '',
+                'max_wind': '',
+                'max_wind_aloft': '',
+                'min_humid': '',
+                'min_temp': '',
+                'min_temp_hour': '',
+                'min_wind': '',
+                'min_wind_aloft': '',
+                'moisture_100hr': '',
+                'moisture_10hr': '',
+                'moisture_1hr': '',
+                'moisture_1khr': '',
+                'moisture_duff': '',
+                'moisture_live': '',
+                'nh3': '',
+                'nox': '',
+                'pm10': '',
+                'pm2.5': 40.0,
+                'rain_days': '',
+                'slope': '',
+                'snow_month': '',
+                'so2': '',
+                'state': '',
+                'sunrise_hour': '',
+                'sunset_hour': '',
+                'type': 'WF',
+                "utc_offset": "-06:00",
+                'voc': ''
+            }
+        ]
+
+        assert len(fires_fields) == len(expected_fires_fields)
+        for i in range(len(fires_fields)):
+            assert fires_fields[i].keys() == expected_fires_fields[i].keys()
+            for k in fires_fields[i]:
+                assert fires_fields[i][k] == expected_fires_fields[i][k], "{} differs".format(k)
+
+        expected_events_fields = {
+            'bigeventid': {
+                'name':'foo-event',
+                'total_area': 598,
+                'total_ch4': None,
+                'total_co': None,
+                'total_co2': None,
+                'total_heat': None,
+                'total_nh3': None,
+                'total_nmhc': None,
+                'total_nox': None,
+                'total_pm10': None,
+                'total_pm2.5': 100, # because pm2.5 wasn't defined for one location
                 'total_so2': None,
                 'total_voc': None
             }